added jwt token and added delete button

+++ backend/app.py
@@ -2,9 +2,19 @@ from flask import Flask, request, jsonify
 from flask_sqlalchemy import SQLAlchemy
 from flask_cors import CORS
 import bcrypt
+from flask_jwt_extended import(
+    JWTManager,
+    create_access_token,
+    jwt_required,
+    get_jwt_identity   
+)
+from datetime import timedelta
 
 app = Flask(__name__)
 CORS(app)
+app.config["JWT_SECRET_KEY"] = "super-secret-key"
+jwt = JWTManager(app)
+app.config["JWT_ACCESS_TOKEN_EXPIRES"] = timedelta(hours=1)
 
 app.config["SQLALCHEMY_DATABASE_URI"] = "mysql+pymysql://root:fdbXASyqhZsQhfnwyjMUyunAiyWPpWij@yamanote.proxy.rlwy.net:13608/railway"
 
@@ -79,9 +89,11 @@ def login():
         data["password"].encode("utf-8"),
         user.password.encode("utf-8")
     ):
+        access_token = create_access_token(identity=str(user.id))
         return jsonify({
             "message": "Login successful",
-            "user_id": user.id
+            "user_id": user.id,
+            "access_token": access_token
         })
 
     return jsonify({"error": "Invalid credentials"}), 401
@@ -89,10 +101,12 @@ def login():
 
 # ---------------- TASK CRUD ----------------
 
-@app.route("/tasks/<int:user_id>", methods=["GET"])
-def get_tasks(user_id):
-    tasks = Task.query.filter_by(user_id=user_id).all()
-
+@app.route("/tasks", methods=["GET"])
+@jwt_required()
+def get_tasks():
+    current_user = int(get_jwt_identity())
+    tasks = Task.query.filter_by(user_id=current_user).all()
+    
     return jsonify([
         {
             "id": t.id,
@@ -105,14 +119,18 @@ def get_tasks(user_id):
 
 
 @app.route("/tasks", methods=["POST"])
+@jwt_required()
 def add_task():
     data = request.json
-
+    if not data.get("title"):
+        return jsonify({"error": "Title is required"}), 400 
+    current_user_id = int(get_jwt_identity())
     task = Task(
         title=data["title"],
         description=data.get("description"),
-        user_id=data["user_id"]
+        user_id=current_user_id
     )
+   
 
     db.session.add(task)
     db.session.commit()
@@ -121,11 +139,17 @@ def add_task():
 
 
 @app.route("/tasks/<int:task_id>", methods=["PUT"])
+@jwt_required()
 def update_task(task_id):
+    current_user_id = int(get_jwt_identity())
+    
     task = Task.query.get(task_id)
 
     if not task:
         return jsonify({"error": "Task not found"}), 404
+    
+    if task.user_id != current_user_id:
+        return jsonify({"error": "Unauthorized"}), 401
 
     data = request.json
     task.title = data.get("title", task.title)
@@ -138,11 +162,17 @@ def update_task(task_id):
 
 
 @app.route("/tasks/<int:task_id>", methods=["DELETE"])
+@jwt_required()
 def delete_task(task_id):
+    current_user_id = int(get_jwt_identity())
+    
     task = Task.query.get(task_id)
 
     if not task:
         return jsonify({"error": "Task not found"}), 404
+    
+    if task.user_id != current_user_id:
+        return jsonify({"error": "Unauthorized"}), 401
 
     db.session.delete(task)
     db.session.commit()
@@ -153,4 +183,4 @@ def delete_task(task_id):
 # ---------------- RUN ----------------
 
 if __name__ == "__main__":
-    app.run(debug=False)
+    app.run(debug=True)
